UPSTREAM: dma-buf/heaps: Correct the types of fd_flags and heap_flags
dma_heap_allocation_data defines the UAPI as follows:

 struct dma_heap_allocation_data {
        __u64 len;
        __u32 fd;
        __u32 fd_flags;
        __u64 heap_flags;
 };

But dma heaps are casting both fd_flags and heap_flags into
unsigned long. This patch makes dma heaps - cma heap and
system heap have consistent types with UAPI.

Acked-by: John Stultz <jstultz@google.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240606020213.49854-1-21cnbao@gmail.com
Git-commit: b9578c49456340ca4d3c7ddbaca054ffc2b51bc1
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[ quic_pbrahma@quicinc.com: Updated qcom heap apis as per new signature. ]

+++ drivers/dma-buf/heaps/qcom_cma_heap.c
@@ -7,7 +7,7 @@
  * Copyright (C) 2012, 2019 Linaro Ltd.
  * Author: <benjamin.gaignard@linaro.org> for ST-Ericsson.
  * Copyright (c) 2020-2021, The Linux Foundation. All rights reserved.
- * Copyright (c) 2023 Qualcomm Innovation Center, Inc. All rights reserved.
+ * Copyright (c) 2023-2024 Qualcomm Innovation Center, Inc. All rights reserved.
  */
 
 #include <linux/cma.h>
@@ -52,8 +52,8 @@ static void cma_heap_free(struct qcom_sg_buffer *buffer)
 /* dmabuf heap CMA operations functions */
 struct dma_buf *cma_heap_allocate(struct dma_heap *heap,
 				  unsigned long len,
-				  unsigned long fd_flags,
-				  unsigned long heap_flags)
+				  u32 fd_flags,
+				  u64 heap_flags)
 {
 	struct cma_heap *cma_heap;
 	struct qcom_sg_buffer *helper_buffer;
